Fixed progress & invites

+++ BusinessLMS/Models/Progress.cs
@@ -7,7 +7,7 @@ namespace BusinessLMS.Models
     {
         public long ProgressId { get; set; }
         public int GoalId { get; set; }
-        public double progress { get; set; }
+        public decimal progress { get; set; }
         public DateTime datetime { get; set; }
     }
 }
